refactor: 전반적 리팩터링
- 사용하지 않는 예외 삭제
- DTO 객체에 final 키워드 추가
- 패키지명 변경
