+++ src/main/java/subway/domain/Line.java
@@ -8,11 +8,11 @@ public class Line {
     private final String name;
     private final String color;
 
-    public Line(String name, String color) {
+    public Line(final String name, final String color) {
         this(null, name, color);
     }
 
-    public Line(Long id, String name, String color) {
+    public Line(final Long id, final String name, final String color) {
         this.id = id;
         this.name = name;
         this.color = color;
