+++ src/main/java/subway/domain/Station.java
@@ -7,12 +7,12 @@ public class Station {
     private final Long id;
     private final String name;
 
-    public Station(Long id, String name) {
+    public Station(final Long id, final String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Station(String name) {
+    public Station(final String name) {
         this(null, name);
     }
