+++ src/main/java/subway/dto/StationResponse.java
@@ -3,10 +3,11 @@ package subway.dto;
 import subway.domain.Station;
 
 public class StationResponse {
-    private Long id;
-    private String name;
 
-    public StationResponse(Long id, String name) {
+    private final Long id;
+    private final String name;
+
+    public StationResponse(final Long id, final String name) {
         this.id = id;
         this.name = name;
     }
