--- src/main/java/subway/exception/IllegalAddStationException.java
@@ -1,8 +1,0 @@
-package subway.exception;
-
-public class IllegalAddStationException extends RuntimeException {
-
-    public IllegalAddStationException(final String message) {
-        super(message);
-    }
-}
