--- src/main/java/subway/exception/LineNotFoundException.java
@@ -1,10 +1,0 @@
-package subway.exception;
-
-public class LineNotFoundException extends RuntimeException {
-
-    private static final String MESSAGE = "호선이 존재하지 않습니다.";
-
-    public LineNotFoundException() {
-        super(MESSAGE);
-    }
-}
