--- src/main/java/subway/exception/StationAlreadyExistException.java
@@ -1,10 +1,0 @@
-package subway.exception;
-
-public class StationAlreadyExistException extends RuntimeException {
-
-    private static final String MESSAGE = "기존에 역이 존재합니다.";
-
-    public StationAlreadyExistException() {
-        super(MESSAGE);
-    }
-}
