+++ src/main/java/subway/presentation/SectionController.java
@@ -1,4 +1,4 @@
-package subway.ui;
+package subway.presentation;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import subway.application.SectionService;
+import subway.service.SectionService;
 import subway.dto.AddOneSectionRequest;
 import subway.dto.AddTwoSectionRequest;
 import java.net.URI;
