+++ src/main/java/subway/service/StationService.java
@@ -1,4 +1,4 @@
-package subway.application;
+package subway.service;
 
 import org.springframework.stereotype.Service;
 import subway.dao.StationDao;
